+++ domains/colour/internal/model/colourpot.go
@@ -4,6 +4,6 @@ import "github.com/google/uuid"
 
 type ColourPot struct {
 	PotID   uuid.UUID
-	VolumeL int
+	VolumeL float64
 	Colour  Colour
 }
