+++ domains/colour/internal/usecase/events/paintaddedtopot.go
@@ -5,10 +5,10 @@ import "github.com/haagor/RGB/domains/colour/internal/model"
 type PaintAddedToPot struct {
 	BaseEvent
 	Colour  model.Colour
-	VolumeL int
+	VolumeL float64
 }
 
-func NewPaintAddedToPot(baseEvent BaseEvent, colour model.Colour, volume int) *PaintAddedToPot {
+func NewPaintAddedToPot(baseEvent BaseEvent, colour model.Colour, volume float64) *PaintAddedToPot {
 	return &PaintAddedToPot{
 		BaseEvent: baseEvent,
 		Colour:    colour,
